refactor: ListNotes moved to its own component
