+++ src/pages/MainLayout/index.ts
@@ -1,0 +1,2 @@
+import { MainLayout } from './MainLayout';
+export { MainLayout };
